feat(web): settings only show up where they can actually be saved
Global settings persist to the primary environment's settings.json, and
the hosted web app never owns a primary connection target. Writes to
those settings were silently discarded there, with no log and no error.

Text generation model moves from General to the Providers page. It now
writes per device through useUpdateEnvironmentSettings, so it works in
the hosted web app once a device is connected.

Source Control drops out of the sidebar, the settings search index and
its own route whenever no primary environment exists. Every control on
that page writes to the primary, so the page had nothing to offer.
