+++ apps/web/src/components/settings/settingsSearch.ts
@@ -16,6 +16,17 @@ export interface SettingsSearchItem {
 }
 
 /**
+ * Sections built entirely on settings that persist to the primary
+ * environment's `settings.json`. They have no backing store where no primary
+ * environment exists (notably the hosted static web app), so the nav, the
+ * search index and the route itself all drop them there rather than showing a
+ * page whose every write is discarded.
+ */
+export const PRIMARY_ONLY_SETTINGS_PATHS: ReadonlySet<SettingsPath> = new Set<SettingsPath>([
+  "/settings/source-control",
+]);
+
+/**
  * Section labels in sidebar order. The sidebar nav and the search-result
  * subtitles both render from this record, so each label exists once.
  */
@@ -147,11 +158,6 @@ export const SETTINGS_SEARCH_ITEMS = [
     to: "/settings/general",
   },
   {
-    id: "text-generation-model",
-    title: "Text generation model",
-    to: "/settings/general",
-  },
-  {
     id: "diagnostics",
     title: "Diagnostics",
     to: "/settings/general",
@@ -204,6 +210,13 @@ export const SETTINGS_SEARCH_ITEMS = [
   {
     id: "providers",
     title: "Providers",
+    to: "/settings/providers",
+  },
+  {
+    // Per-device: each environment carries its own selection in its
+    // `settings.json`, so the row renders inside the selected device's panel.
+    id: "text-generation-model",
+    title: "Text generation model",
     to: "/settings/providers",
   },
   {
